requirements.txt
selenium>=4.0.0
pytest>=7.0.0
mysql-connector-python>=8.0.0
paramiko>=2.9.0
pytest-html>=3.0.0
webdriver-manager>=3.8.0
python-dotenv>=0.19.0